# docs/swagger.json
{
    "schemes": [
        "http"
    ],
    "swagger": "2.0",
    "info": {
        "description": "This is a server for storing NASA apod data \u0026 images.",
        "title": "APOD Serevr",
        "contact": {},
        "version": "1.0"
    },
    "host": "localhost:8080",
    "basePath": "/",
    "paths": {
        "/get": {
            "get": {
                "description": "Gets all recorded Data",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "imageInfo"
                ],
                "summary": "Get All Images Info",
                "responses": {
                    "200": {
                        "description": "image infos",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/model.ImageInfoResponse"
                            }
                        }
                    },
                    "404": {
                        "description": "error message",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/get/{date}": {
            "get": {
                "description": "Gets an image info on specific date",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "imageInfo"
                ],
                "summary": "Get Image Info",
                "responses": {
                    "200": {
                        "description": "image info",
                        "schema": {
                            "$ref": "#/definitions/model.ImageInfoResponse"
                        }
                    },
                    "404": {
                        "description": "error message",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        }
    },
    "definitions": {
        "model.ImageInfoResponse": {
            "type": "object",
            "properties": {
                "copyright": {
                    "description": "Copyright of the image",
                    "type": "string",
                    "example": "\nBrennan Gilmore\n"
                },
                "date": {
                    "description": "Date wnen image became apod",
                    "type": "string",
                    "example": "2024-10-14"
                },
                "explanation": {
                    "description": "Explanation of the image",
                    "type": "string",
                    "example": " Go outside at sunset tonight and see a comet!  C/2023 A3 (Tsuchinshan–ATLAS) has become visible..."
                },
                "hdurl": {
                    "description": "HD URL of the image (link to nasa website)",
                    "type": "string",
                    "example": "https://apod.nasa.gov/apod/image/2410/CometA3Dc_Gilmore_1080.jpg"
                },
                "media_type": {
                    "description": "MediaType of the date provided",
                    "type": "string",
                    "example": "image"
                },
                "proxy_url": {
                    "description": "URL of the stored image (proxy image)",
                    "type": "string",
                    "example": "http://localhost:8009/2024-10-14.jpg"
                },
                "service_version": {
                    "description": "Nasa service version",
                    "type": "string",
                    "example": "v1"
                },
                "title": {
                    "description": "Title of the image provided",
                    "type": "string",
                    "example": "Comet Tsuchinshan-ATLAS Over the Lincoln Memorial"
                },
                "url": {
                    "description": "URL of the image (link to nasa website)",
                    "type": "string",
                    "example": "https://apod.nasa.gov/apod/image/2410/CometA3Dc_Gilmore_1080.jpg"
                }
            }
        }
    }
}